// ocyt_frontend/src/interface.ts
export interface Diseno {
  region: null | string;
  pais: Pais;
  title?: string;
  residente: Residente;
  solicitante: string;
  locarno: number | null;
  pais_Nombre: PaisNombre;
  dimension: Dimension;
  inventor: null | string;
  ciudad: null | string;
  solicitud_ANONIMIZADA: string;
  solicitud?: number;
  ano_Presentacion: string;
  fecha_de_Presentacion: null | string;
}

export enum Dimension {
  DiseñoIndustrial2D = "Diseño Industrial 2D",
  DiseñoIndustrial3D = "Diseño Industrial 3D",
}

export enum Pais {
  Ar = "AR",
  At = "AT",
  Au = "AU",
  Aw = "AW",
  Be = "BE",
  Bg = "BG",
  Bm = "BM",
  Br = "BR",
  CA = "CA",
  CN = "CN",
  Ch = "CH",
  Cl = "CL",
  Co = "CO",
  Cz = "CZ",
  De = "DE",
  Dk = "DK",
  Do = "DO",
  Ec = "EC",
  Es = "ES",
  Fi = "FI",
  Fr = "FR",
  GB = "GB",
  Gr = "GR",
  Gt = "GT",
  Hk = "HK",
  Hu = "HU",
  IL = "IL",
  Ie = "IE",
  In = "IN",
  It = "IT",
  Jp = "JP",
  Kp = "KP",
  Kr = "KR",
  Ky = "KY",
  Li = "LI",
  Lu = "LU",
  MT = "MT",
  MX = "MX",
  Mu = "MU",
  My = "MY",
  Nl = "NL",
  No = "NO",
  PE = "PE",
  Pa = "PA",
  Ph = "PH",
  Pl = "PL",
  Pt = "PT",
  QA = "QA",
  Ru = "RU",
  SE = "SE",
  Sg = "SG",
  Si = "SI",
  Sk = "SK",
  Th = "TH",
  Tr = "TR",
  Tt = "TT",
  Tw = "TW",
  Ua = "UA",
  Us = "US",
  Uy = "UY",
  VG = "VG",
  Ve = "VE",
  Vn = "VN",
  Za = "ZA",
}

export enum PaisNombre {
  Alemania = "Alemania",
  Argentina = "Argentina",
  Aruba = "Aruba",
  Australia = "Australia",
  Austria = "Austria",
  Bermuda = "Bermuda",
  Brasil = "Brasil",
  Bulgaria = "Bulgaria",
  Bélgica = "Bélgica",
  Canadá = "Canadá",
  Chequia = "Chequia",
  Chile = "Chile",
  China = "China",
  Colombia = "Colombia",
  Dinamarca = "Dinamarca",
  Ecuador = "Ecuador",
  Eslovaquia = "Eslovaquia",
  Eslovenia = "Eslovenia",
  España = "España",
  EstadosUnidosDeAmérica = "Estados Unidos de América",
  FederaciónDeRusia = "Federación de Rusia",
  Filipinas = "Filipinas",
  Finlandia = "Finlandia",
  Francia = "Francia",
  Grecia = "Grecia",
  Guatemala = "Guatemala",
  HongKong = "Hong Kong",
  Hungría = "Hungría",
  India = "India",
  Irlanda = "Irlanda",
  IslasCaimán = "Islas Caimán",
  IslasVírgenesBritánicas = "Islas Vírgenes Británicas",
  Israel = "Israel",
  Italia = "Italia",
  Japón = "Japón",
  Liechtenstein = "Liechtenstein",
  Luxemburgo = "Luxemburgo",
  Malasia = "Malasia",
  Malta = "Malta",
  Mauricio = "Mauricio",
  México = "México",
  Noruega = "Noruega",
  Panamá = "Panamá",
  PaísesBajos = "Países Bajos",
  Perú = "Perú",
  Polonia = "Polonia",
  Portugal = "Portugal",
  Qatar = "Qatar",
  ReinoUnidoDeGranBretañaEIrlandaDelNorte = "Reino Unido de Gran Bretaña e Irlanda del Norte",
  RepúblicaDeCorea = "República de Corea",
  RepúblicaDominicana = "República Dominicana",
  RepúblicaPopularDemocráticaDeCorea = "República Popular Democrática de Corea",
  Singapur = "Singapur",
  Sudáfrica = "Sudáfrica",
  Suecia = "Suecia",
  Suiza = "Suiza",
  Tailandia = "Tailandia",
  Taiwán = "Taiwán",
  TrinidadYTobago = "Trinidad y Tobago",
  Turquía = "Turquía",
  Ucrania = "Ucrania",
  Uruguay = "Uruguay",
  Venezuela = "Venezuela",
  Vietnam = "Vietnam",
}

export enum Residente {
  NoResidente = "No residente",
  Residente = "Residente",
}